refactored folder structure, map prey to 1 policy and predator to 1 policy
